+++ app/Models/FeeType.php
@@ -65,7 +65,7 @@ class FeeType extends Model
     protected function amount(): Attribute
     {
         return Attribute::make(
-            get: fn (int $value) => $value / 100,
+            get: fn (int $value) => (float) ($value / 100),
             set: fn (float $value) => (int) round($value * 100),
         );
     }
